Курсач по пикпо

+++ book-author-service/src/main/java/com/example/library/repository/GenreRepository.java
@@ -1,9 +1,9 @@
 package com.example.library.repository;
-import com.example.library.entity.GenreEntity;
+import com.example.library.model.dao.GenreEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GenreRepository extends JpaRepository<GenreEntity, Long> {
+public interface GenreRepository extends JpaRepository<GenreEntity, Integer> {
 }
